student_management/requirements.txt
Django==3.2 
djangorestframework 
djongo 
